app/requirements.txt
# Core backend
django==4.1.1
djangorestframework==3.14.0
django-filter==22.1
django-extensions==3.2.1

# Webserver
gunicorn==20.1.0

# Auto Generated Swagger Doc
drf_spectacular==0.24.2